app/supabase.py
import os
from dotenv import load_dotenv
from supabase import create_client, Client
from fastapi import Request, HTTPException

# Load environment variables from .env file
load_dotenv()

# Supabase configuration
SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")

# Initialize Supabase client
supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)

async def get_authenticated_user(request: Request):
    """Verify Supabase JWT token from request headers"""
    auth_header = request.headers.get("Authorization")
    if not auth_header or not auth_header.startswith("Bearer "):
        raise HTTPException(status_code=401, detail="Invalid authorization header")

    token = auth_header.split(" ")[1]  # Extract the token
    try:
        user = supabase.auth.get_user(token)  # Validate the token with Supabase
    except Exception as e:
        raise HTTPException(status_code=401, detail="Invalid or expired token")

    if not user:
        raise HTTPException(status_code=401, detail="Invalid or expired token")
    
    return user