requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
langgraph>=0.0.15
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=0.19.0
supabase
aiohttp>=3.8.0